Seeded the databasewith hardcoded data and removed it from app.js file
