+++ views/index.html
@@ -53,23 +53,19 @@
 
 <div id="saleitems">
 
-
-	<div class="col-lg-4 col-sm-6">
-		<div class="thumbnail">
-			<div style="background-image:url(/images/hp.jpg)" class="bg-image"></div>
-		</div>
-	</div>
-<div class="col-lg-4 col-sm-6">
-	<div class="thumbnail">
-		<div style="background-image:url(/images/mattress.jpg)" class="bg-image"></div>
-	</div>
-</div>
-<div class="col-lg-4 col-sm-6">
-	<div class="thumbnail">
-		<div style="background-image:url(/images/samsung.jpg)" class="bg-image"></div>
-	</div>
+<!-- <div class="form-group">
+  <label class="col-md-4 control-label" for="name">Product Name</label>
+  <div class="col-md-4">
+  <input id="name" name="name" type="text" placeholder="" class="form-control input-md" required="">
+  </div>
 </div>
 
+<div class="form-group">
+  <label class="col-md-4 control-label" for="productDescription">Description:</label>
+  <div class="col-md-4">
+  <input id="textinput" name="productDescription" type="text" placeholder="" class="form-control input-md">
+  </div>
+</div> -->
 
 </div>
